Reorganize controller concerns
