+++ app/controllers/concerns/archangel/controllers/authorizable_concern.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Archangel
+  ##
+  # Controller concerns
+  #
+  module Controllers
+    ##
+    # Controller authorize concern
+    #
+    module AuthorizableConcern
+      extend ActiveSupport::Concern
+
+      included do
+        after_action :verify_authorized
+      end
+    end
+  end
+end
