+++ app/controllers/concerns/archangel/controllers/skip_authorizable_concern.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Archangel
+  ##
+  # Controller concerns
+  #
+  module Controllers
+    ##
+    # Controller skip authorize concern
+    #
+    module SkipAuthorizableConcern
+      extend ActiveSupport::Concern
+
+      included do
+        after_action :skip_authorization
+      end
+    end
+  end
+end
